cli/pyproject.toml
[tool.poetry]
name = "fastchargeapi-cli"
version = "0.0.2"
description = "Official CLI for managing apps and account on FastchargeAPI.com"
authors = ["FastchargeAPI <fastchargeapi@gmail.com>"]
license = "GNU v3.0"
homepage = "https://fastchargeapi.com"
repository = "https://github.com/FastchargeAPI/fastchargeapi-cli"
packages = [{include = "fastchargeapi_cli"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
gql = {extras = ["all"], version = "^3.4.0"}
colorama = "^0.4.6"
click-aliases = "^1.0.1"
blessings = "^1.7"
sgqlc = "^16.1"
emoji = "^2.2.0"
aws-requests-auth = "^0.4.3"
flask-cors = "^3.0.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tomli = "^2.0.1"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}

[tool.poetry.scripts]
fastcharge = 'fastchargeapi_cli.main:fastcharge'
fastapi = 'fastchargeapi_cli.main:fastapi'

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
turms = "0.4.2"
pydantic = "^1.10.6"
pyright = "^1.1.299"
boto3 = "^1.26.114"
pytest-xdist = "^3.2.1"
pytest = "^7.3.1"
isort = "^5.12.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]

[tool.coverage.run]
omit = [
    "fastchargeapi_cli/config.py",
]
